refactor(website): scope iframe URL listener with Mount instead of Subscription
The example detail iframe needs to know when the embedded app navigates,
which it signals via window.postMessage. Previously a Subscription listened
on window globally and gated itself on the route being ExampleDetail. The
real cause of the listener is the iframe existing in the rendered tree, so
Mount.defineStream on the iframe is a more direct expression: the listener's
lifetime tracks the element, and event.source === iframe.contentWindow
scopes filtering to the specific frame rather than guessing from the route.
