+++ packages/website/src/subscription/index.ts
@@ -1,6 +1,5 @@
 export * as ActiveSection from './activeSection'
 export * as AiHeading from './aiHeading'
-export * as ExampleUrl from './exampleUrl'
 export * as SearchShortcut from './searchShortcut'
 export * as SystemTheme from './systemTheme'
 export * as ViewportWidth from './viewportWidth'
